Allow empty metric families in Prometheus exposition
This adds in a test for gravelgateway#17, which was a bug where a metric family
with a descriptor, but without any metrics, would cause an invalid parse.

To fix this, we update the parser to allow for empty metric families, and
update the tests to ensure that we can parse the empty metric family.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>