Adds support for untyped Prometheus metrics.
When the type of a metric is not known, the Prometheus exposition format uses the type "untyped". This differs from the OpenMetrics specification, which uses "unknown". This can be seen in the official Prometheus text parser:
https://github.com/prometheus/prometheus/blob/e2b9cfeeeb1efbe299c386226622a5064f23f377/model/textparse/promparse.go#L317

This change also adds a test case drawn from the latest version of the Prometheus Node Exporter. (1.7.0)